Outsource root routing from app module
The root routing has been added to its own file, which is imported in
the app routing module.

Furthermore, the Angular Material version has been set to stick to
beta2.
