--- frontend/src/app/app-routing/app-routing.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@NgModule({
-  imports: [
-    CommonModule
-  ],
-  declarations: []
-})
-export class AppRoutingModule { }
